requirements.txt
pywebio
openai
pyPDF2
google-cloud-storage